update evm state extraction
